use auxiliary functionality from testtools package
