+++ test/compression_structures_test.go
@@ -7,43 +7,12 @@ import (
 	"github.com/wal-g/wal-g/internal"
 	"github.com/wal-g/wal-g/internal/compression"
 	"github.com/wal-g/wal-g/internal/compression/lz4"
+	"github.com/wal-g/wal-g/testtools"
 	"io"
 	"io/ioutil"
 	"math/rand"
 	"testing"
 )
-
-var MockCloseError = errors.New("mock close: close error")
-var MockReadError = errors.New("mock reader: read error")
-var MockWriteError = errors.New("mock writer: write error")
-
-type BufCloser struct {
-	*bytes.Buffer
-	err bool
-}
-
-func (w *BufCloser) Close() error {
-	if w.err {
-		return MockCloseError
-	}
-	return nil
-}
-
-type ErrorWriteCloser struct{}
-
-func (ew ErrorWriteCloser) Write(p []byte) (int, error) {
-	return -1, MockWriteError
-}
-
-func (ew ErrorWriteCloser) Close() error {
-	return MockCloseError
-}
-
-type ErrorReader struct{}
-
-func (er ErrorReader) Read(p []byte) (int, error) {
-	return -1, MockReadError
-}
 
 func GetLz4Compressor() compression.Compressor {
 	return compression.Compressors[lz4.AlgorithmName]
@@ -59,7 +28,7 @@ var tests = []struct {
 
 func TestCascadeFileCloser(t *testing.T) {
 	for _, tt := range tests {
-		b := &BufCloser{bytes.NewBufferString(tt.testString), false}
+		b := &testtools.BufCloser{bytes.NewBufferString(tt.testString), false}
 		lz := &internal.CascadeWriteCloser{
 			WriteCloser: GetLz4Compressor().NewWriter(b),
 			Underlying:  b,
@@ -78,7 +47,7 @@ func TestCascadeFileCloser(t *testing.T) {
 		err = lz.Close()
 		assert.NoErrorf(t, err, "compress: CascadeWriteCloser expected `<nil>` but got %v", err)
 
-		b.err = true
+		b.Err = true
 
 		err = lz.Close()
 		assert.Errorf(t, err, "compress: Underlying writer expected to close with error but got `<nil>`")
@@ -87,7 +56,7 @@ func TestCascadeFileCloser(t *testing.T) {
 }
 
 func TestCascadeFileCloserError(t *testing.T) {
-	mock := &ErrorWriteCloser{}
+	mock := &testtools.ErrorWriteCloser{}
 	lz := &internal.CascadeWriteCloser{
 		WriteCloser: GetLz4Compressor().NewWriter(mock),
 		Underlying:  mock,
@@ -102,11 +71,11 @@ func TestCascadeFileCloserError(t *testing.T) {
 
 func TestCompressAndEncrypt(t *testing.T) {
 	for _, tt := range tests {
-		in := &BufCloser{bytes.NewBufferString(tt.testString), false}
+		in := &testtools.BufCloser{bytes.NewBufferString(tt.testString), false}
 		compressor := GetLz4Compressor()
 		compressed := internal.CompressAndEncrypt(in, compressor, nil)
 
-		decompressed := &BufCloser{&bytes.Buffer{}, false}
+		decompressed := &testtools.BufCloser{&bytes.Buffer{}, false}
 		decompressor := compression.GetDecompressorByCompressor(compressor)
 		err := decompressor.Decompress(decompressed, compressed)
 		if err != nil {
@@ -122,12 +91,12 @@ func TestCompressAndEncryptBigChunk(t *testing.T) {
 	L := 1024 * 1024 // 1Mb
 	b := make([]byte, L)
 	rand.Read(b)
-	in := &BufCloser{bytes.NewBuffer(b), false}
+	in := &testtools.BufCloser{bytes.NewBuffer(b), false}
 
 	compressor := GetLz4Compressor()
 	compressed := internal.CompressAndEncrypt(in, compressor, nil)
 
-	decompressed := &BufCloser{&bytes.Buffer{}, false}
+	decompressed := &testtools.BufCloser{&bytes.Buffer{}, false}
 	decompressor := compression.GetDecompressorByCompressor(compressor)
 	err := decompressor.Decompress(decompressed, compressed)
 	if err != nil {
@@ -160,11 +129,11 @@ func testCompressAndEncryptErrorPropagation(compressor compression.Compressor, t
 	L := 1 << 20
 	b := make([]byte, L)
 	rand.Read(b)
-	in := &BufCloser{bytes.NewBuffer(b), false}
+	in := &testtools.BufCloser{bytes.NewBuffer(b), false}
 
 	compressed := internal.CompressAndEncrypt(in, compressor, nil)
 
-	decompressed := &BufCloser{&bytes.Buffer{}, false}
+	decompressed := &testtools.BufCloser{&bytes.Buffer{}, false}
 	decompressor := compression.GetDecompressorByCompressor(compressor)
 	err := decompressor.Decompress(decompressed, &DelayedErrorReader{compressed, L})
 	assert.Errorf(t, err, "%v did not propagate error of the buffer", compressor.FileExtension())
@@ -178,7 +147,7 @@ func TestCompressAndEncryptErrorPropagation(t *testing.T) {
 
 func TestCompressAndEncryptError(t *testing.T) {
 	compressor := GetLz4Compressor()
-	compressed := internal.CompressAndEncrypt(&ErrorReader{}, compressor, nil)
+	compressed := internal.CompressAndEncrypt(&testtools.ErrorReader{}, compressor, nil)
 
 	_, err := ioutil.ReadAll(compressed)
 	assert.Errorf(t, err, "compress: CompressingPipeWriter expected error but got `<nil>`")
